// front/src/app/api-client/model/shopping-list-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShoppingItemDTO } from './shopping-item-dto';


/**
 * Liste de courses de la famille : articles générés depuis les repas et articles ajoutés à la main
 */
export interface ShoppingListDTO { 
    /**
     * Début de la période générée, absent si la liste n\'a jamais été générée
     */
    dateDebut?: string | null;
    dateFin?: string | null;
    genereLe?: string | null;
    /**
     * Articles, par ordre alphabétique (les lignes retirées n\'y figurent pas)
     */
    articles?: Array<ShoppingItemDTO>;
}